refactor: remove unused schema v1 code
