+++ warehouse/schema/schema_v2.go
@@ -3,21 +3,59 @@ package schema
 import (
 	"context"
 	"fmt"
+	"regexp"
 	"slices"
 	"sync"
 	"time"
 
 	"github.com/samber/lo"
 
+	"github.com/rudderlabs/rudder-go-kit/config"
+
 	"github.com/rudderlabs/rudder-go-kit/logger"
 	"github.com/rudderlabs/rudder-go-kit/stats"
 	obskit "github.com/rudderlabs/rudder-observability-kit/go/labels"
 	"github.com/rudderlabs/rudder-server/jsonrs"
 	"github.com/rudderlabs/rudder-server/utils/timeutil"
+	"github.com/rudderlabs/rudder-server/warehouse/integrations/middleware/sqlquerywrapper"
 	"github.com/rudderlabs/rudder-server/warehouse/internal/model"
 	"github.com/rudderlabs/rudder-server/warehouse/internal/repo"
+	"github.com/rudderlabs/rudder-server/warehouse/logfield"
 	whutils "github.com/rudderlabs/rudder-server/warehouse/utils"
 )
+
+// deprecatedColumnsRegex
+// This regex is used to identify deprecated columns in the warehouse
+// Example: abc-deprecated-dba626a7-406a-4757-b3e0-3875559c5840
+var deprecatedColumnsRegex = regexp.MustCompile(
+	`.*-deprecated-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$`,
+)
+
+type schemaRepo interface {
+	GetForNamespace(ctx context.Context, sourceID, destID, namespace string) (model.WHSchema, error)
+	Insert(ctx context.Context, whSchema *model.WHSchema) (int64, error)
+}
+
+type stagingFileRepo interface {
+	GetSchemasByIDs(ctx context.Context, ids []int64) ([]model.Schema, error)
+}
+
+type fetchSchemaRepo interface {
+	FetchSchema(ctx context.Context) (model.Schema, error)
+}
+
+type Handler interface {
+	SyncRemoteSchema(ctx context.Context, fetchSchemaRepo fetchSchemaRepo, uploadID int64) (bool, error)
+	IsWarehouseSchemaEmpty(ctx context.Context) bool
+	GetTableSchemaInWarehouse(ctx context.Context, tableName string) model.TableSchema
+	UpdateLocalSchema(ctx context.Context, updatedSchema model.Schema) error
+	UpdateWarehouseTableSchema(ctx context.Context, tableName string, tableSchema model.TableSchema) error
+	GetColumnsCountInWarehouseSchema(ctx context.Context, tableName string) (int, error)
+	ConsolidateStagingFilesUsingLocalSchema(ctx context.Context, stagingFiles []*model.StagingFile) (model.Schema, error)
+	UpdateLocalSchemaWithWarehouse(ctx context.Context) error
+	TableSchemaDiff(ctx context.Context, tableName string, tableSchema model.TableSchema) (whutils.TableSchemaDiff, error)
+	FetchSchemaFromWarehouse(ctx context.Context, repo fetchSchemaRepo) error
+}
 
 type schemaV2 struct {
 	stats struct {
@@ -37,19 +75,33 @@ type schemaV2 struct {
 	cachedSchemaMu                   sync.RWMutex
 }
 
-func newSchemaV2(v1 *schema, warehouse model.Warehouse, log logger.Logger, ttlInMinutes time.Duration, fetchSchemaRepo fetchSchemaRepo) *schemaV2 {
+func New(
+	db *sqlquerywrapper.DB,
+	warehouse model.Warehouse,
+	conf *config.Config,
+	logger logger.Logger,
+	statsFactory stats.Stats,
+	fetchSchemaRepo fetchSchemaRepo,
+) Handler {
+	ttlInMinutes := conf.GetDurationVar(720, time.Minute, "Warehouse.schemaTTLInMinutes")
 	v2 := &schemaV2{
 		warehouse:                        warehouse,
-		log:                              log,
+		log:                              logger.Child("schema_v2"),
 		ttlInMinutes:                     ttlInMinutes,
-		schemaRepo:                       v1.schemaRepo,
-		stagingFilesSchemaPaginationSize: v1.stagingFilesSchemaPaginationSize,
-		stagingFileRepo:                  v1.stagingFileRepo,
+		schemaRepo:                       repo.NewWHSchemas(db),
+		stagingFilesSchemaPaginationSize: conf.GetInt("Warehouse.stagingFilesSchemaPaginationSize", 100),
+		stagingFileRepo:                  repo.NewStagingFiles(db),
 		fetchSchemaRepo:                  fetchSchemaRepo,
-		enableIDResolution:               v1.enableIDResolution,
+		enableIDResolution:               conf.GetBool("Warehouse.enableIDResolution", false),
 		now:                              timeutil.Now,
 	}
-	v2.stats.schemaSize = v1.stats.schemaSize
+	v2.stats.schemaSize = statsFactory.NewTaggedStat("warehouse_schema_size", stats.HistogramType, stats.Tags{
+		"module":        "warehouse",
+		"workspaceId":   warehouse.WorkspaceID,
+		"destType":      warehouse.Destination.DestinationDefinition.Name,
+		"sourceId":      warehouse.Source.ID,
+		"destinationId": warehouse.Destination.ID,
+	})
 	return v2
 }
 
@@ -215,4 +267,188 @@ func (sh *schemaV2) getSchema(ctx context.Context) (model.Schema, error) {
 	sh.cachedSchema = whSchema.Schema
 	sh.cacheExpiry = whSchema.ExpiresAt
 	return sh.cachedSchema, nil
+}
+
+// consolidateStagingSchemas merges multiple schemas into one
+// Prefer the type of the first schema, If the type is text, prefer text
+func consolidateStagingSchemas(consolidatedSchema model.Schema, schemas []model.Schema) model.Schema {
+	for _, schema := range schemas {
+		for tableName, columnMap := range schema {
+			if _, ok := consolidatedSchema[tableName]; !ok {
+				consolidatedSchema[tableName] = model.TableSchema{}
+			}
+			for columnName, columnType := range columnMap {
+				if columnType == model.TextDataType {
+					consolidatedSchema[tableName][columnName] = model.TextDataType
+					continue
+				}
+
+				if _, ok := consolidatedSchema[tableName][columnName]; !ok {
+					consolidatedSchema[tableName][columnName] = columnType
+				}
+			}
+		}
+	}
+	return consolidatedSchema
+}
+
+// consolidateWarehouseSchema overwrites the consolidatedSchema with the schemaInWarehouse
+// Prefer the type of the schemaInWarehouse, If the type is text, prefer text
+func consolidateWarehouseSchema(consolidatedSchema, warehouseSchema model.Schema) model.Schema {
+	for tableName, columnMap := range warehouseSchema {
+		if _, ok := consolidatedSchema[tableName]; !ok {
+			continue
+		}
+
+		for columnName, columnType := range columnMap {
+			if _, ok := consolidatedSchema[tableName][columnName]; !ok {
+				continue
+			}
+
+			var (
+				consolidatedSchemaType = consolidatedSchema[tableName][columnName]
+				warehouseSchemaType    = columnType
+			)
+
+			if consolidatedSchemaType == model.TextDataType && warehouseSchemaType == model.StringDataType {
+				continue
+			}
+
+			consolidatedSchema[tableName][columnName] = columnType
+		}
+	}
+
+	return consolidatedSchema
+}
+
+// overrideUsersWithIdentifiesSchema overrides the users table with the identifies table
+// users(id) <-> identifies(user_id)
+// Removes the user_id column from the users table
+func overrideUsersWithIdentifiesSchema(consolidatedSchema model.Schema, warehouseType string, warehouseSchema model.Schema) model.Schema {
+	var (
+		usersTable      = whutils.ToProviderCase(warehouseType, whutils.UsersTable)
+		identifiesTable = whutils.ToProviderCase(warehouseType, whutils.IdentifiesTable)
+		userIDColumn    = whutils.ToProviderCase(warehouseType, "user_id")
+		IDColumn        = whutils.ToProviderCase(warehouseType, "id")
+	)
+	if _, ok := consolidatedSchema[usersTable]; !ok {
+		return consolidatedSchema
+	}
+	if _, ok := consolidatedSchema[identifiesTable]; !ok {
+		return consolidatedSchema
+	}
+
+	for k, v := range consolidatedSchema[identifiesTable] {
+		consolidatedSchema[usersTable][k] = v
+	}
+	for k, v := range warehouseSchema[usersTable] {
+		if _, ok := warehouseSchema[identifiesTable][k]; !ok {
+			consolidatedSchema[usersTable][k] = v
+			consolidatedSchema[identifiesTable][k] = v
+		}
+	}
+	consolidatedSchema[usersTable][IDColumn] = consolidatedSchema[identifiesTable][userIDColumn]
+	delete(consolidatedSchema[usersTable], userIDColumn)
+	return consolidatedSchema
+}
+
+// enhanceDiscardsSchema adds the discards table to the schema
+// For bq, adds the loaded_at column to be segment compatible
+func enhanceDiscardsSchema(consolidatedSchema model.Schema, warehouseType string) model.Schema {
+	discards := model.TableSchema{}
+
+	for colName, colType := range whutils.DiscardsSchema {
+		discards[whutils.ToProviderCase(warehouseType, colName)] = colType
+	}
+
+	if warehouseType == whutils.BQ {
+		discards[whutils.ToProviderCase(warehouseType, "loaded_at")] = "datetime"
+	}
+
+	consolidatedSchema[whutils.ToProviderCase(warehouseType, whutils.DiscardsTable)] = discards
+	return consolidatedSchema
+}
+
+// enhanceSchemaWithIDResolution adds the merge rules and mappings table to the schema if IDResolution is enabled
+func enhanceSchemaWithIDResolution(consolidatedSchema model.Schema, isIDResolutionEnabled bool, warehouseType string) model.Schema {
+	if !isIDResolutionEnabled {
+		return consolidatedSchema
+	}
+	var (
+		mergeRulesTable = whutils.ToProviderCase(warehouseType, whutils.IdentityMergeRulesTable)
+		mappingsTable   = whutils.ToProviderCase(warehouseType, whutils.IdentityMappingsTable)
+	)
+	if _, ok := consolidatedSchema[mergeRulesTable]; ok {
+		consolidatedSchema[mergeRulesTable] = model.TableSchema{
+			whutils.ToProviderCase(warehouseType, "merge_property_1_type"):  "string",
+			whutils.ToProviderCase(warehouseType, "merge_property_1_value"): "string",
+			whutils.ToProviderCase(warehouseType, "merge_property_2_type"):  "string",
+			whutils.ToProviderCase(warehouseType, "merge_property_2_value"): "string",
+		}
+		consolidatedSchema[mappingsTable] = model.TableSchema{
+			whutils.ToProviderCase(warehouseType, "merge_property_type"):  "string",
+			whutils.ToProviderCase(warehouseType, "merge_property_value"): "string",
+			whutils.ToProviderCase(warehouseType, "rudder_id"):            "string",
+			whutils.ToProviderCase(warehouseType, "updated_at"):           "datetime",
+		}
+	}
+	return consolidatedSchema
+}
+
+func removeDeprecatedColumns(schema model.Schema, warehouse model.Warehouse, log logger.Logger) {
+	for tableName, columnMap := range schema {
+		for columnName := range columnMap {
+			if deprecatedColumnsRegex.MatchString(columnName) {
+				log.Debugw("skipping deprecated column",
+					logfield.SourceID, warehouse.Source.ID,
+					logfield.DestinationID, warehouse.Destination.ID,
+					logfield.DestinationType, warehouse.Destination.DestinationDefinition.Name,
+					logfield.WorkspaceID, warehouse.WorkspaceID,
+					logfield.Namespace, warehouse.Namespace,
+					logfield.TableName, tableName,
+					logfield.ColumnName, columnName,
+				)
+				delete(schema[tableName], columnName)
+			}
+		}
+	}
+}
+
+func tableSchemaDiff(tableName string, schemaMap model.Schema, tableSchema model.TableSchema) whutils.TableSchemaDiff {
+	diff := whutils.TableSchemaDiff{
+		ColumnMap:        make(model.TableSchema),
+		UpdatedSchema:    make(model.TableSchema),
+		AlteredColumnMap: make(model.TableSchema),
+	}
+
+	currentTableSchema, ok := schemaMap[tableName]
+
+	if !ok {
+		if len(tableSchema) == 0 {
+			return diff
+		}
+		diff.Exists = true
+		diff.TableToBeCreated = true
+		diff.ColumnMap = tableSchema
+		diff.UpdatedSchema = tableSchema
+		return diff
+	}
+
+	for columnName, columnType := range currentTableSchema {
+		diff.UpdatedSchema[columnName] = columnType
+	}
+
+	diff.ColumnMap = make(model.TableSchema)
+	for columnName, columnType := range tableSchema {
+		if _, ok := currentTableSchema[columnName]; !ok {
+			diff.ColumnMap[columnName] = columnType
+			diff.UpdatedSchema[columnName] = columnType
+			diff.Exists = true
+		} else if columnType == model.TextDataType && currentTableSchema[columnName] == model.StringDataType {
+			diff.AlteredColumnMap[columnName] = columnType
+			diff.UpdatedSchema[columnName] = columnType
+			diff.Exists = true
+		}
+	}
+	return diff
 }
